requirements.txt
mediapipe==0.10.9
numpy
opencv-python
dollarpy
MCRcon
python-dotenv